Made install path detection registry independent
Sorted out some duplicated files.
